Moving stuff around and setting up for rendering
Added a package for the data management stuff, added an ObjectWrapper
class to contain extra information about each object (such as whether or
not it is visible, name, etc.) Used the camerax sample solution for
full-screen. (In our case, recreating the view on rotation is just not
acceptable, since that involves re-initializing the TLEs, this looks
smoother though).

Added some test code for rendering.